gdb: Clean up inferior list when reconnecting to new target
When connected to a target and running an inferior, if the inferior
forks then a new, inferior is created to track the child
process. Usually the second inferior (the child process) will be
automatically cleaned up (removed) by GDB once the child has completed.
However, if we connect to a new target then the second inferior is left
around, and we can even end up using the second inferior in order to run
a program on the new target, for example in this session:

   (gdb) target extended-remote :2347
   (gdb) file /path/to/exe
   (gdb) set remote exec-file /path/to/exe
   (gdb) set detach-on-fork off
   (gdb) break breakpt
   (gdb) run
   # ... hits breakpoint
   (gdb) info inferiors
     Num  Description       Executable
   * 1    process 15406     /path/to/exe
     2    process 15407     /path/to/exe
   (gdb) kill
   Kill the program being debugged? (y or n) y
   (gdb) info inferiors
     Num  Description       Executable
   * 1    <null>            /path/to/exe
     2    process 15433     /path/to/exe
   (gdb) target extended-remote :2348
   (gdb) file /path/to/exe
   (gdb) set remote exec-file /path/to/exe
   (gdb) run
   # ... hits breakpoint
   (gdb) info inferiors
     Num  Description       Executable
     1    <null>            /path/to/exe
   * 2    process 15408     /path/to/exe
     3    process 15409     /path/to/exe
   (gdb)

Notice how after connecting to the second extended-remote target, and
then running the test program we now have inferiors 1, 2, and 3.
There's nothing really _wrong_ with this, but a better experience would,
I think, be to have only inferiors 1 and 2 in the new target's session.

The issue here is that in target.c:dispose_inferior GDB uses
switch_to_thread, which also switches the current inferior.  This leaves
the last inferior in the list selected after target.c:target_preopen has
completed.

We could change target.c:dispose_inferior to ensure the selected
inferior is not left changed, however, in the above example, I think
that, even if the user manually selected inferior 2 _before_ connecting
to the new target, the best behaviour would still be for GDB to switch
back to inferior 1.  This seems to suggest that the right thing to do is
have GDB _choose_ a new inferior as part of target_preopen, after which
is can call prune_inferiors to remove transient inferiors from the list.

An existing test is extended to cover this behaviour.

gdb/ChangeLog:

	* target.c (find_first_killed_inferior): New function.
	(target_preopen): Use find_first_killed_inferior to reset selected
	inferior, and prune killed inferior as appropriate.

gdb/testsuite/ChangeLog:

	* gdb.server/extended-remote-restart.exp: Extend.