WIP: Try to fix overread when fetching component
The fix is guarded with the environment variable APB_FIX.

This contains loads of debug output.

Disabling pretty printers is never going to fly, especially as even
with the overread, printing virtual inherited baseclasses works fine.

This is thanks to a hack/work-around/feature in
valops.c:do_search_struct_field that reloads the baseclass information
if we spot that the offset to the baseclass is outside the bounds of
the value we have to hand.

Current strategy is to avoid value_from_component and create a new
method value_from_component_2 to use instead.