objcopy: Don't assume section names start with DOT when removing relocs
When calling --remove-section=.rela.text to remove relocations from
the .text section we had an assuption that section names begin with a
'.'.  If a section didn't, then the relocation section would be named,
for example .relaFOO (for a section named 'FOO'), in which case we
would fail to remove the relocations.

Fixed in this commit.

binutils/ChangeLog:

	* objcopoy.c (handle_remove_section_option): Don't assume that
	section names begin with '.'.
	* testsuite/binutils-all/remove-relocs-07.d: New file.
	* testsuite/binutils-all/remove-relocs-07.s: New file.
	* testsuite/binutils-all/remove-relocs-08.d: New file.