WIP:gdb: Improve target-description reggroup creation
The target descriptions mechanism looses track of some of the regieter
sets.  This is because some sets are pre-defined, like 'general' and
'float'.

When we do reggroup_find, if there are no target specific reggroups
then we look in the default list.  But, as soon as we add a custom
target specific regset we create a target specific list, and use that.

So, imagine a target description with register in these sets:

general, float, custom, general

When we call tdesc_use_registers we first look for the regset
'general' - this is found in the default list, then we look for
'float', also in the default list, then we look for 'custom'.  This is
not in the default list, and so we create a new list, and add custom
too it.

Finally, we look for 'general', don't find it, create a __NEW__ regset
called general, and add this to the target specific list.

The gdbarch's list of reggroups now contains 'custom' and 'general'.
And the 'general' is NOT the same general as was in the default list.

This patch is a partial solution, it takes care of ensuring that all
reggroups do end up on the gdbarch list, but doesn't handle the mess
with duplicate reggroups of the same name being created.

It's worth noting that having multiple reggroups of the same name
probably doesn't matter, but might be a nice cleanup.