gdb/regcache: When saving, ignore registers that can't be read
If we fail to read a register during reg_buffer::save then this should
cause GDB to terminate as it currently does, instead we should mark
the register as unavailable, and then when we try to restore the
registers, that unsaved register will not be restored.

This was observed on a RISC-V remote target which announces that a
particular register is available in the target description, but then
gives an error whenever the register is read.

Yes, this is a bad target, but, it feels like GDB should handle this
better than it currently does (crash).

gdb/ChangeLog:

	* regcache.c (reg_buffer::save): When saving the current register
	state, ignore registers that can't be read.