gdb: Enable stdin on exception in execute_gdb_command
This is a revised version of the patch that was submitted here:

  https://sourceware.org/ml/gdb-patches/2018-09/msg00884.html

This patch attempts to address PR gdb/23718 by re-enabling stdin
whenever an exception is caught during gdb.execute().

When Python gdb.execute() is called, an exception could occur (e.g. the
target disappearing), which is then converted into a Python exception.
If stdin was disabled and an exception is caught, it is not necessarily
re-enabled, because the exception doesn't propagate to the top level of
the event loop, whose catch block would enable it, and we may also not
hit normal_stop (), which would enable stdin. The result is that when
execution of a Python script completes, GDB does not prompt or accept
input, and is effectively hung.

This change rectifies this issue by re-enabling stdin in the CATCH block
of execute_gdb_command, prior to converting the exception to a Python
exception.

gdb/ChangeLog:

	PR gdb/23718
	* gdb/python/python.c (execute_gdb_command): Call
	async_enable_stdin in CATCH block.

gdb/testsuite/ChangeLog:

        PR gdb/23718
	* gdb.server/server-kill-python.c: New file.
	* gdb.server/server-kill-python.exp: New file.