WIP: Find start of earliest baseclass
This commit finds the start of the earliest baseclass for our type.
This offset can be negative when dealing with virtual base classes
(though it can also be positive too in that case).

The idea I have right now it to combine this offset with the embedded
offset, and then maybe have "offset" be an offset from the embedded
offset??

|AAAAABBBBBCCCCCCC-----------------------------|
      |    |
      |   ptr
      '----' Negative offset to base class.