gdb: Move initialised check out of each langauges pretty-printers
Currently each extension langauage's pretty printer checks that the
extension langauge is initialised, and then prepares a copy of the
value for printing.

This means that there's some duplicate code, and if that code is wrong
(as it currently is) then the bug must be fixed in multiple places.

This commit is part of a series aiming to fix the incorrect
pretty-printer code.  The first step, move the initialised check up
into the general extension langauge wrapper code.

gdb/ChangeLog:

	* extension.c (apply_ext_lang_val_pretty_printer): Check extension
	langauge is initialised before calling pretty printer method.
	* guile/scm-pretty-print.c (gdbscm_apply_val_pretty_printer):
	Assert that the extension language is initialised, and remove
	check.
	* python/py-prettyprint.c (gdbpy_apply_val_pretty_printer): Likewise.