gdb: Move value preparation out of each languages pretty printer
Each extension language currently creates a copy of the value to print
using the same code sequence.  The problem is that this code sequence
is wrong.

As part of a patch series to fix this code, this commit moves all of
this standard boiler plate code out of each extension language's
pretty printer method, and places it in the generic extension.c
function instead.

gdb/ChangeLog:

	* extension-priv.h (struct extension_language_ops): Change
	signature of apply_val_pretty_printer method.
	* extension.c (apply_ext_lang_val_pretty_printer): Build value
	before calling language specific pretty printer.  Update call to
	language pretty printer.
	* guile/guile-internal.h: Update signature of value pretty printer
	method.
	* guile/scm-pretty-print.c (gdbscm_apply_val_pretty_printer):
	Update signature of value pretty printer method, and remove value
	preparation code.
	* python/py-prettyprint.c (gdbpy_apply_val_pretty_printer):
	Likewise.
	* python/python-internal.h: Update signature of value pretty printer
	method.