gdb: Don't flush line wrap buffer before adding newline
A bug was introduced in commit:

    commit cbe5657196d0d3acbeca39973f93f333ecedacda
    Date:   Mon Sep 3 22:56:33 2018 -0600

        Add output styles to gdb

As GDB produces output, the output is stored in a temporary buffer.
In this way GDB can spot when a line wrap occurs, and can insert a
newline before flushing the contents of the temporary buffer out.

The bug was that the temporary buffer was being flushed out _before_
the newline was added to the output stream, this resulted in incorrect
line wrapping.  Here is an example from how GDB announces that a
breakpoint has been hit, first with the bug:

Breakpoint 1, function () at /a/long/pa
    th/to/the/source/file.c:123

Now with the bug fixed (this matches the behaviour before the
offending commit):

Breakpoint 1, function ()
    at /a/long/path/to/the/source/file.c:123

There are some other issues with line wrapping in current HEAD of GDB,
but these are caused by an unrelated bug.

gdb/ChangeLog:

	* utils.c (fputs_maybe_filtered): Don't flush the line wrap buffer
	before inserting a newline for line-wrapping.

gdb/testsuite/ChangeLog:

	* gdb.base/line-wrapping.c: New file.
	* gdb.base/line-wrapping.exp: New file.