gdb: Make use of show function for set/show variables in target.c
Using a separate show function allows for internationalisation of the
text used for the show message, and reduces our dependence on
deprecated_show_value_hack.

I've mostly kept the show message as they were, except for:

  show may-write-registers

Where I changed the message from:

  Permission to write into registers is on.

to

  Permission to write into target registers is on.

This is inline with may-write-memory, may-insert-breakpoints,
may-insert-tracepoints, may-insert-fast-tracepoints, may-interrupt,
all of which use the word "target" in their output message.

I've also updated the set/show help text for may-write-registers to
include the word "target".