gdb: Make use of show function for set/show variables in i386-tdep.c
Using a separate show function allows for internationalisation of the
text used for the show message and reduces our dependence on
deprecated_show_value_hack.

The text in the new show functions is copied from the message that is
currently auto-generated, there should be no user visible changes
after this commit.