gdb: Make use of show function for set/show variables in ser-tcp.c
Use a separate show function for set/show commands in ser-tcp.c, this
allows for internationalisation of the text used in the show message,
and reduces our dependence on deprecated_show_value_hack.

The text to display in the new show functions was copied from the
previously auto-generated text.

I also added some help text to the set/show tcp auto-retry setting.