gdb: Make use of show function for set/show variables in fork-child.c
Using a separate show function allows for internationalisation of the
text used for the show message.  Also split the description of the
setting out from the 'set' string so that it can be displayed as part
of the help text for both 'set' and 'show'.

The new show function special cases for the empty string to show a
more helpful message.

gdb/ChangeLog:

	* fork-child.c (show_exec_wrapper): New function.
	(_initialize_fork_child): Split description of setting into
	separate field, and pass show function.