gdb: Add an is_declaration field to each symbol
This commit is in preparation for a later commit, there should be no
user visible change after this commit.

Track if a symbol is a declaration or not.  At one point we could
possibly figure this out based on the LOC_UNRESOLVED address class of
the symbol, but, for Fortran we mark some symbols as LOC_UNRESOLVED
even when the DWARF supplies an address, this is because the address
supplied by the DWARF is actually wrong.  For details look in
dwarf2read.c and look for references to gFortran bug #40040.  I have
confirmed that current versions of gFortran still have the issue
mentioned in that bug report.

I considered two possible solutions to this problem, one was to add a
new address class, something like:
    LOC_UNRESOLVED_BUT_IS_STILL_A_DEFINITION
that I could use specifically for marking these "broken" Fortran
symbols.  My concern here is that every place LOC_UNRESOLVED is
mentioned would need to be possibly edited to also handle the new
address class, and any future change that mentions LOC_UNRESOLVED
would be a possible source of Fortran bugs due to not handling the new
address class correctly.

So, the solution I took was to add a new single bit flag to the symbol
to track if the symbol is a declaration or not.  There are already
some 1 bit flags in the symbol object, and on x86-64 GNU/Linux (at
least for me) adding one additional flag didn't increase the size of
the symbol object at all.

gdb/ChangeLog:

	* dwarf2read.c (new_symbol): Mark symbol as declaration when
	appropriate.
	* symtab.h (struct symbol): Add 'is_declaration' flag.
	(SYMBOL_IS_DECLARATION): Define.