gdb/fortran: Include module variables in 'info variables' output
Due to a work around in dwarf2read.c for gFortran issue #40040,
non-constant module variables don't show up in 'info variables'
output.

The reason is that issue #40040 means that the debug symbol definition
for these module variables can sometimes have the incorrect location
information attached to it.  The work around in GDB is to ignore the
location information from the debug information, and mark these
symbols as LOC_UNRESOLVED, GDB will then fall back to using the
non-debug symbols in order to figure out the location of these
symbols.

The problem with this is that GDB uses LOC_UNRESOLVED to indicate a
variable declaration, and doesn't include such symbols in its output.

In an earlier commit I extended each symbol to now explicitly track if
it is marked as a declaration in the DWARF, so GDB can now use this
information rather than LOC_UNRESOLVED to exclude declarations from
the 'info variables' output.

You might think that we could just switch from checking:

  SYMBOL_CLASS (sym) != LOC_UNRESOLVED

to checking:

  !SYMBOL_IS_DECLARATION (sym)

However, this is not good enough, often in C or C++ the declaration is
_also_ the definition, in which case the symbol will be marked as a
declaration, but will _not_ be of address class LOC_UNRESOLVED.  The
correct condition then is that we shouldn't show declarations that are
of class LOC_UNRESOLVED.  The preserves the existing C/C++ behaviour
in every case I've tried so far, and also fixes the gFortran issues
that I was seeing.

gdb/ChangeLog:

	* dwarf2read.c (new_symbol): Mark symbol as declaration when
	appropriate.
	* symtab.h (struct symbol): Add 'is_declaration' flag.
	(SYMBOL_IS_DECLARATION): Define.

gdb/testsuite/ChangeLog:

	* gdb.fortran/module.exp: Extend with 'info variables' test.