gdb: Add a show function for user settings in linux-thread-db.c
Using a separate show function allows for internationalisation of the
text used for the show message, and reduces our dependence on
deprecated_show_value_hack.

The text to display in the new show functions was mostly copied from
the previously auto-generated text, however, I spotted a typo in the
'show libthread-db-search-path' text, we had 'or' instead of 'for'.
So I've fixed this in the new show function, and in the original show
help text.