gdb: Add show function for user settings in x86-nat.c
Using a separate show function allows for internationalisation of the
text used for the show message, and reduces our dependence on
deprecated_show_value_hack.

The text to display in the new show functions was copied from the
previously auto-generated text.

There should be no user visible changes after this commit.