gdb: Use show functions for user settings in python and guile
Use a separate show function for the set/show commands in the python
and guile sub-directories.  This allows for internationalisation of the
text used in the show message, and reduces our dependence on
deprecated_show_value_hack.

The text to display in the new show functions was copied from the
previously auto-generated text.