WIP: Test for reconnecting to target with multiple inferiors
I had an old branch that tried to fix the "connect to target with
multiple inferiors" issue, however Pedro's multi-target work has now
fixed this issue.

The only thing from that branch that I think has any value is the
test, which I believe is fuller than anything we currently have in
this area.

Once Pedro's multi-target work has been merged then this test should
be added.

I have updated this test so that it should pass when run on top of
Pedro's branch.