gdb/testsuite: Add gdb_test_ext helper function
Add a new helper function gdb_test_ext.  This is a cross between
gdb_test and gdb_test_multiple.  Here is an example of its use taken
from gdb.base/store.exp:

    gdb_test_ext "print l" "${prefix}; print old l, expecting ${l}" {
	" = ${l}" {
	    pass $gdb_test_ext_name
	}
	" = <optimized out>" {
	    unsupported $gdb_test_ext_name
	    set supported 0
	}
    }

The first argument is the GDB command to send, and the second argument
is the test name.  The third argument is a list of patterns and action
bodies.

The patterns don't have '-re' in front because they are not full
patterns - they will be modified by having $gdb_prompt appended, just
like in gdb_test.

Inside action bodies you have access to a variable $gdb_test_ext_name,
this is set up by the call to gdb_test_ext and contains a copy of the
test name, this can be used by the calls to pass/fail/etc and avoids
having to setup the test name in a separate variable.

gdb/testsuite/ChangeLog:

	* gdb.base/store.exp (check_set): Use gdb_test_ext instead of
	gdb_test_multiple.
	(up_set): Likewise.
	* lib/gdb.exp (gdb_test_ext): New procedure.