gdb/python: New block parameter for gdb.lookup_static_symbol
NOTE: This patch was abandoned after this discussion on the mailing
      list:

      https://sourceware.org/ml/gdb-patches/2019-11/msg00014.html

Adds a new block parameter to gdb.lookup_static_symbol.  Currently
gdb.lookup_static_symbol first looks in the global static scope of the
current block for a matching symbol, and if non is found it then looks
in all other global static scopes.

The new block parameter for gdb.lookup_static_symbol allows the user
to search the global static scope for some other block rather than the
current block if they so desire.

If the block parameter is no given, or is given as None, then GDB will
search using the current block first (this matches the CLI behaviour).

I added the block parameter to the middle of the parameter list, so
the parameter order is 'symbol block domain', this matches the order
of the other symbol lookup function that takes a block 'lookup_symbol'
which also has the order 'symbol block domain'.  I think changing the
API like this is fine as the 'gdb.lookup_static_symbol' has not been
in any previous GDB release, so there should be no existing user
scripts that use this function.

The documentation for gdb.lookup_static_symbol has been reordered so
that the block parameter can be explained early on, and I've added
some tests.

gdb/ChangeLog:

	* python/py-symbol.c (gdbpy_lookup_static_symbol): Accept and
	handle a new block parameter.

gdb/testsuite/ChangeLog:

	* gdb.python/py-symbol.exp: Add tests for passing a block
	parameter to gdb.lookup_static_symbol.

gdb/doc/ChangeLog:

	* python.texi (Symbols In Python): Rewrite documentation for
	gdb.lookup_static_symbol, including adding a description of the
	new block parameter.

Change-Id: I132a7f0934ba029c9f32058e9341a6c7cb975606