gdb/fortran: Allow array index accesses to complex number components
Add a new mechanism for Fortran to allow the components of a complex
number to be accessed using array syntax notation.  So, instead of a
user writing this:

  (gdb) p $_creal (complex_var)

they can instead write this:

  (gdb) p complex_var (1)

And similarly, instead of this:

  (gdb) p $_cimag (complex_var)

they can use:

  (gdb) p complex_var (2)

There is not reason why one approach is better then the other, this is
simply an easy feature to add that some users might find easier to
remember or figure out.  Especially when printing complex_var (in this
case) prints a result that looks very much like a two element array:

  (gdb) p complex_var
  $1 = (1,2)
  (gdb) p array_var
  $2 = (1,2,3)

So it might seem natural to a user that the components of complex_var
can be accessed using the name array notation.

gdb/ChangeLog:

	* eval.c (evaluate_subexp_standard): Support array index access
	for components of complex numbers for Fortran.

gdb/doc/ChangeLog:

	* gdb.texinfo (Fortran Complex Numbers): New node describing array
	subscript accesses to complex numbers.

gdb/testsuite/ChangeLog:

	* gdb.fortran/complex.exp: Rewrite to use the new proc.
	(test_complex_value): New proc.

Change-Id: I94389b82608adf57fcef86d21b40b007691bb119