Add new configure options to allow static libraries to be selected
This commit modifies the configure scripts of various projects to add
some new options that allow users to force the selection of a static
library even when the shared library is available.

Lets take as an example the GDB project and libexpat library.
Currently GDB depends on libexpat, and if the shared library is
installed on the users machine GDB will detect this and link against
it.

If the shared library is not available, but the static library is,
then GDB will link against this.

However, if both libraries are available then GDB will always select
the shared library first, and there's currently no way to force
selection of the static library.

This makes sense on Linux based systems and is undoubtedly the correct
choice, however, in some cases a user might want to build GDB using a
static libexpat, but removing the shared library version of libexpat
from a machine could be tricky (if we want to not break other things
on the machine).

This commit adds the configure flag --with-libexpat-type=TYPE, where
type is either auto (the default), static, or shared.  If the user
selects either static or shared then the configure process will ignore
the other type of library, even if it is present on the machine.

As this change is actually made in one of the underlying configure
macros (AC_LIB_LINKFLAGS_BODY) many libraries gain a new flag of this
type.

Change-Id: I727ddcc75a8bafedb633196ec45b37e519d7b838

config/ChangeLog:

	* lib-link.m4 (AC_LIB_LINKFLAGS_BODY): Add new
	--with-libXXX-type=... option.  Use this to guide the selection of
	either a shared library or a static library.

binutils/ChangeLog:

	* configure: Regenerate.

gdb/ChangeLog:

	* configure: Regenerate.

Change-Id: I01806583a0655d5f99ba53889f9a38631b88af95