gdb: Print frame address in more cases
Change GDB to display the frame address for all inline frames except
the inner most frame.  This issue was discussed briefly on the mailing
list here:

   https://sourceware.org/ml/gdb-patches/2019-12/msg01015.html

Consider this stack setup from the test gdb.dwarf2/dw2-inline-many-frames.exp:

   #11     #10    #9     #8     #7     #6     #5     #4     #3     #2     #1     #0
   main -> aaa -> bbb -> ccc -> ddd -> eee -> fff -> ggg -> hhh -> iii -> jjj -> kkk
   \_______________________/    \________/    \______________________/    \________/
      Inline sequence #1          Normal          Inline sequence #2        Normal

When we stop in GDB and backtrace the stack is displayed like
this (I've replaced the file and line information with "..."):

  (gdb) bt
  #0  kkk () at ...
  #1  0x0000000000400494 in jjj () at ...
  #2  0x00000000004004b1 in iii () at ...
  #3  hhh () at ...
  #4  ggg () at ...
  #5  fff () at ...
  #6  0x0000000000400536 in eee () at ...
  #7  0x0000000000400519 in ddd () at ...
  #8  0x00000000004004f3 in ccc () at ...
  #9  bbb () at ...
  #10 aaa () at ...
  #11 main () at ...
  (gdb)

Notice that many frames are missing an address, take the sequence for
frames #6 to #2.  Frame #6 is a non-inline frame and has an address.
Frame #6 calls to frame #5 which is also non-inline.  Frames #4, #3,
and #2 are all inline within frame #5, and of these only frame #2 has
an address displayed.

This lack of address can be confusing, there's no clear indication why
the addresses are missing, a user needs to understand that the missing
address implies an inline frame, and can then infer the address based
on the address of frame #2.

After this patch the backtrace is now show like this:

  #0  kkk () at ...
  #1  0x0000000000400494 in jjj () at ...
  #2  0x00000000004004b1 in iii () at ...
  #3  0x00000000004004b1 in hhh () at ...
  #4  0x00000000004004b1 in ggg () at ...
  #5  0x00000000004004b1 in fff () at ...
  #6  0x0000000000400536 in eee () at ...
  #7  0x0000000000400519 in ddd () at ...
  #8  0x00000000004004f3 in ccc () at ...
  #9  0x00000000004004f3 in bbb () at ...
  #10 0x00000000004004f3 in aaa () at ...
  #11 0x00000000004004f3 in main () at ...

Now all frames (except #0) have an address.  Frames that are inline,
and previously lacked an address will now have the same address as the
inner-most inline frame in the sequence, so for the inline sequence #5
to #2, all frames have the same address as frame #2.  For the sequence
frame #11 to #8, all have the address of frame #8.

The duplicated address might also be confusing, though (personally) I
think it is slightly more obvious from the duplicated addresses that
the frames are inlined, however, if people strongly disagree and
prefer the no-address layout we could make this feature switchable.

One further location where the no address can crop up, is when the
user switches frame, previously we would see:

  (gdb) frame 3
  #3  hhh () at ...
  115	  return iii () + 1;
  (gdb)

After the patch we see:

  (gdb) frame 3
  #3  0x00000000004004b1 in hhh () at ...
  115	  return iii () + 1;
  (gdb)

There were a small number of tests that needed to have their expected
results updated, and I added a small section to the documentation to
help users understand what duplicate addresses in the backtrace might
mean.

gdb/ChangeLog:

	* stack.c (frame_show_address): Show the frame address for all but
	the inner-most inlined frame.

gdb/testsuite/ChangeLog:

	* gdb.dwarf2/dw2-inline-many-frames.exp: Update expected results.
	* gdb.dwarf2/dw2-inline-param.exp: Likewise.
	* gdb.guile/scm-frame-inline.exp: Likewise.
	* gdb.opt/inline-cmds.exp: Likewise.
	* gdb.python/py-frame-inline.exp: Likewise.

gdb/doc/ChangeLog:

	* gdb.texinfo (Backtrace): Mention duplicate addresses in the
	backtrace.